+++ changelog.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
+## [unreleased]
+
+- Replace --custom option with optional named arguments
+
 ## [0.5.0] - 2022-11-27
 
-- Add flag --custom for coloring by regexp
+- Add option --custom for coloring by regexp
 - Use gregoryv/cmdline
 
 ## [0.4.0] - 2022-11-05